requirements.txt
Django==3.2
djangorestframework==3.12.4
django-filter==2.4.0
dj-database-url==0.5.0
gunicorn==20.1.0